고생하셨습니다

+++ src/main/java/com/iot/pay/payVO.java
@@ -7,19 +7,20 @@ public class payVO {
 	private int rv_id;
 	private String user_id;
 	private String pm_method_code;
-	private int pm_price;
+	private double pm_price;
 	private Date pm_date;
-	private int pm_price_plus;
+	private double pm_price_plus;
 	private int use_point;
 	private int discount;
 	private int card_grant_num;
 	private String pm_confirm;
+	
 	public payVO() {
+	
 	}
 
-
-	public payVO(int pm_num, int rv_id, String user_id, String pm_method_code, int pm_price, Date pm_date,
-			int pm_price_plus, int use_point, int discount, int card_grant_num, String pm_confirm) {
+	public payVO(int pm_num, int rv_id, String user_id, String pm_method_code, double pm_price, Date pm_date,
+			double pm_price_plus, int use_point, int discount, int card_grant_num, String pm_confirm) {
 		super();
 		this.pm_num = pm_num;
 		this.rv_id = rv_id;
@@ -33,7 +34,6 @@ public class payVO {
 		this.card_grant_num = card_grant_num;
 		this.pm_confirm = pm_confirm;
 	}
-
 
 	public int getPm_num() {
 		return pm_num;
@@ -59,22 +59,19 @@ public class payVO {
 		this.user_id = user_id;
 	}
 
-
 	public String getPm_method_code() {
 		return pm_method_code;
 	}
-
 
 	public void setPm_method_code(String pm_method_code) {
 		this.pm_method_code = pm_method_code;
 	}
 
-
-	public int getPm_price() {
+	public double getPm_price() {
 		return pm_price;
 	}
 
-	public void setPm_price(int pm_price) {
+	public void setPm_price(double pm_price) {
 		this.pm_price = pm_price;
 	}
 
@@ -86,11 +83,11 @@ public class payVO {
 		this.pm_date = pm_date;
 	}
 
-	public int getPm_price_plus() {
+	public double getPm_price_plus() {
 		return pm_price_plus;
 	}
 
-	public void setPm_price_plus(int pm_price_plus) {
+	public void setPm_price_plus(double pm_price_plus) {
 		this.pm_price_plus = pm_price_plus;
 	}
